requirements.txt
pandas>=1.5.0
numpy==1.26.4
pandas-ta==0.3.14b
xgboost==1.7.6
scikit-learn>=1.3.0
schedule>=1.2.0
python-dotenv>=1.0.0
joblib>=1.3.0
oandapyV20==0.7.2
setuptools>=60.0.0